refactor(core): promote pkg/refcount to internal/core/refcount (no behavior change)
