+++ internal/core/refcount/refcount_test.go
