+++ internal/core/refcount/refcount_windows.go
